Сделал реализацию моделей User, Post, Написал отдельный компонент QueryBuilder. Создал отдельный глобальный конфиг для хранения данных о подключение к БД и отдельный обработчик для получения данных из глобального конфига.

+++ app/Models/Post.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Database\QueryBuilder;
 use DateTime;
 use PDO;
 
@@ -13,35 +14,51 @@ class Post
     private string $title;
     private string $description;
     private string $datePublished;
-    private array $imagePost;
+    private string $imagePost;
     private int $isActive;
 
-    private $db;
+    private QueryBuilder $db;
     private $result = null;
 
-    public function __construct(string $title = "", string $description = "", array $imagePost = [], bool $isActive = true)
+    public function __construct(QueryBuilder $pdo, string $title = "", string $description = "", string $imagePost = "", bool $isActive = true)
     {
         $this->title = $title;
         $this->description = $description;
         $this->imagePost = $imagePost;
         $this->isActive = $isActive;
-        $this->db = getPDO();
+        $this->db = $pdo;
     }
 
-    public function addPost(string $title, string $description, string $imagePost)
+    public function addPost(string $title, string $description, array $imagePost)
     {
         if (empty($title) || empty($description)) {
             echo "Поле не может быть пустым";
             return;
         }
 
-        $stmt = $this->db->prepare("INSERT INTO `posts` (title, description, image_post) VALUES (?, ?, ?)");
-        $ok = $stmt->execute([$title, $description, $imagePost]);
+        if (!isset($imagePost)) {
+            echo "Выберите картинку!";
+            return;
+        }
 
-        $this->id = $this->db->lastInsertId();
+        $imagePost = $this->uploadImage($imagePost);
+        // dd($imagePost);
+
+        if (is_null($imagePost)) return;
+
+        $ok = $this->db->insert(table: "posts", columns: [
+            "title" => $title,
+            "description" => $description,
+            "image_post" => $imagePost
+        ]);
+
+        // $stmt = $this->db->prepare("INSERT INTO `posts` (title, description, image_post) VALUES (?, ?, ?)");
+        // $ok = $stmt->execute([$title, $description, $imagePost]);
+
+        // $this->id = $this->db->lastInsertId();
 
         if ($ok) {
-            $this->id = $this->db->lastInsertId();
+            $this->id = $this->db->getLastID();
             $newPost = $this->getPostById($this->id);
             $this->id = $newPost->id;
             $this->title = $newPost->title;
@@ -58,17 +75,36 @@ class Post
         return false;
     }
 
-    public function editPost(int $id, string $title = "", string $description = "", string $imagePost = "")
+    public function editPost(int $id, string $title = "", string $description = "", array $imagePost = [])
     {
-        if (!$this->getPostById($id)) {
-            echo "Данный пост не существует!";
+        $currentPost = $this->getPostById($id);
+        if (!$currentPost) {
             return;
         }
 
-        $stmt = $this->db->prepare("UPDATE `posts` SET `title` = ?, `description` = ?, `image_post` = ? WHERE `id` = ?");
-        $stmt->execute([$title, $description, $imagePost, $id]);
+        if (empty($title) && empty($description) && empty($imagePost)) {
+            echo "Все поля не могут быть пустыми. Измените значение хотя бы в одном поле!";
+            return;
+        }
 
-        if ($stmt->rowCount() > 0) {
+        if (!empty($imagePost)) {
+            $imagePost = $this->uploadImage($imagePost);
+            $this->deleteImage($id);
+        } else {
+            $imagePost = $currentPost->image_post;
+        }
+
+        $result = $this->db->update(table: "posts", operator: "=", bindValues: [
+            "id" => $id,
+            "title" => $title,
+            "description" => $description,
+            "image_post" => $imagePost
+        ], where: ["id" => $id]);
+
+        // $stmt = $this->db->prepare("UPDATE `posts` SET `title` = ?, `description` = ?, `image_post` = ? WHERE `id` = ?");
+        // $stmt->execute([$title, $description, $imagePost, $id]);
+
+        if ($result > 0) {
             echo "Пост успешно изменён";
             return true;
         }
@@ -79,14 +115,16 @@ class Post
     public function deletePost(int $id)
     {
         if (!$this->getPostById($id)) {
-            echo "Данный пост не существует!";
             return;
         }
 
-        $stmt = $this->db->prepare("DELETE FROM `posts` WHERE id = ?");
-        $stmt->execute([$id]);
+        $result = $this->db->delete(table: "posts", id: $id);
+        $this->deleteImage(id: $id);
 
-        if ($stmt->rowCount() > 0) {
+        // $stmt = $this->db->prepare("DELETE FROM `posts` WHERE id = ?");
+        // $stmt->execute([$id]);
+
+        if ($result > 0) {
             echo "Пост успешно удалён";
             return true;
         }
@@ -96,20 +134,28 @@ class Post
     }
     public function getAllPosts()
     {
-        $stmt = $this->db->prepare("SELECT * FROM `posts`");
-        $stmt->execute();
-        $this->result = $stmt->fetchAll(PDO::FETCH_OBJ);
+        $this->result = $this->db->getAll(table: "posts")->result();
+        // $stmt = $this->db->prepare("SELECT * FROM `posts`");
+        // $stmt->execute();
+        // $this->result = $stmt->fetchAll(PDO::FETCH_OBJ);
         return $this;
     }
 
     public function getPostById(int $id)
     {
-        $stmt = $this->db->prepare("SELECT * FROM `posts` WHERE id = ?");
-        $stmt->execute([$id]);
-        $this->result = $stmt->fetch(PDO::FETCH_OBJ);
-        return $this->result;
+        $this->result = $this->db->getByCondition(table: "posts", operator: "=", columns: ["*"], where: ["id" => $id])->getOneResult();
+
+        if (is_null($this->result())) {
+            echo "Данный пост не существует";
+            return;
+        }
+
+        // $stmt = $this->db->prepare("SELECT * FROM `posts` WHERE id = ?");
+        // $stmt->execute([$id]);
+        // $this->result = $stmt->fetch(PDO::FETCH_OBJ);
+        return $this->result();
     }
-    public function uploadImage(array $image, ?int $idPost = null)
+    public function uploadImage(array $image)
     {
         $arrayMIMEType = ["image/png", "image/jpeg"];
         $nameFile = pathinfo($image["name"], PATHINFO_FILENAME);
@@ -129,73 +175,65 @@ class Post
             if (!move_uploaded_file($tmpName, $targetFile)) {
                 echo "Произошла ошибка при загрузке файла на сервер";
                 return;
+            } else {
+                return $relativePathImage;
             }
-            $stmt = $this->db->prepare("UPDATE `posts` SET `image_post` = ? WHERE id = ?");
-            $stmt->execute([$relativePathImage, $idPost]);
-        }
-
-        if ($stmt->rowCount() > 0) {
-            echo "Картинка загружена успешно";
-            return;
-        } else {
-            unlink($targetFile);
-            echo "Непредвиденная ошибка при загрузке картинки";
-            return false;
         }
     }
 
     public function deleteImage(int $id)
     {
-        
-        $stmt = $this->db->prepare("SELECT `image_post` FROM `posts` WHERE id = ?");
-        $stmt->execute([$id]);
-        $this->result = $stmt->fetch(PDO::FETCH_OBJ);
+        if(!$this->getPostById($id)) {
+            return;
+        }
+
+        $this->result = $this->db->getByCondition(table: "posts", operator: "=", columns: ["image_post"], where: ["id" => $id])->getOneResult();
+
+        // $stmt = $this->db->prepare("SELECT `image_post` FROM `posts` WHERE id = ?");
+        // $stmt->execute([$id]);
+        // $this->result = $stmt->fetch(PDO::FETCH_OBJ);
 
         $targetDirectory = dirname($_SERVER["DOCUMENT_ROOT"]);
         $targetFile = $this->result()->image_post;
-        dd($targetDirectory);
-        dd($targetFile);
 
         $file = $targetDirectory . $targetFile;
 
-        if(isset($this->result)) {
-            if(file_exists($file)) {
+        if (isset($this->result)) {
+            if (file_exists($file)) {
                 unlink($file);
+                return;
             }
 
-            if(!file_exists($file)){
-                $stmt = $this->db->prepare("UPDATE `posts` SET `image_post` = ? WHERE id = ?");
-                $stmt->execute(["", $id]);
-            }
+            // if (!file_exists($file)) {
+            //     $stmt = $this->db->prepare("UPDATE `posts` SET `image_post` = ? WHERE id = ?");
+            //     $stmt->execute(["", $id]);
+            // }
 
-            if($stmt->rowCount() > 0) {
-                echo "Файл успешно удалён с сервера и БД";
-                return true;
-            }
+            // if ($stmt->rowCount() > 0) {
+            //     echo "Файл успешно удалён с сервера и БД";
+            //     return true;
+            // }
         }
 
         echo "Произошла непредвиденная ошибка при удалении файла";
         return false;
     }
 
-    // public function setFolderUploadFile(string $srcDirectory = "", string $nameFolder = "")
-    // {
-    //     if (!empty($srcDirectory)) {
-    //         // return __DIR__ . ""
-    //     }
-    // }
-
     public function archive(int $id, int $isActive = 0)
     {
         if (!$this->getPostById($id)) {
-            echo "Данный пост не существует!";
             return;
         }
 
-        $stmt = $this->db->prepare("UPDATE `posts` SET `is_active` = ? WHERE `id` = ?");
-        $stmt->execute([$isActive, $id]);
+        $result = $this->db->update(table: "posts", operator: "=", bindValues: [
+            "id" => $id,
+            "is_active" => $isActive
+        ], where: ["id" => $id]);
 
-        if ($stmt->rowCount() > 0) {
+        // $stmt = $this->db->prepare("UPDATE `posts` SET `is_active` = ? WHERE `id` = ?");
+        // $stmt->execute([$isActive, $id]);
+
+        if ($result > 0) {
             echo "Пост помещён в архив";
             return true;
         }
@@ -207,19 +245,23 @@ class Post
     public function unarchive(int $id, int $isActive = 1)
     {
         if (!$this->getPostById($id)) {
-            echo "Данный пост не существует!";
             return;
         }
 
-        $stmt = $this->db->prepare("UPDATE `posts` SET `is_active` = ? WHERE `id` = ?");
-        $stmt->execute([$isActive, $id]);
+        $result = $this->db->update(table:"posts", operator:"=", bindValues:[
+            "id" => $id,
+            "is_active" => $isActive
+        ], where:["id" => $id]);
 
-        if ($stmt->rowCount() > 0) {
+        // $stmt = $this->db->prepare("UPDATE `posts` SET `is_active` = ? WHERE `id` = ?");
+        // $stmt->execute([$isActive, $id]);
+
+        if ($result > 0) {
             echo "Пост активен вновь";
             return true;
         }
 
-        echo "Ошибка. Не удалось поместить пост из архив";
+        echo "Ошибка. Не удалось переместить пост из архив";
         return false;
     }
